requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
ipython>=7.0.0
plotly>=5.18.0
scipy>=1.11.0
seaborn>=0.13.0
umap-learn>=0.5.0
